+++ R/3-objectives.R
@@ -76,6 +76,49 @@ objective.maxRDA <- function(codon, ARGS){
   return(varExplained)
 }
 
+#' Calculate Fitness of Binary Vector
+#'
+#' This objective function seeks to maximize the percent of the
+#'  total variance present in the matrix z that can be explained
+#'  by the constraining amalgams.
+#'
+#' @param codon A binary vector.
+#' @param ARGS Handled by \code{\link{prepareArgs}}.
+#' @export
+objective.maxRDA2 <- function(codon, ARGS){
+
+  W <- do.call(ARGS$weights, list(codon, n.amalgams = ARGS$n.amalgams))
+  A <- ARGS$x %*% W
+
+  # Don't allow zeros!
+  if(any(A == 0)){
+    return(-1000000)
+  }
+
+  # Maximize variance in Z explained by AMALG
+  tryCatch({
+
+    if(ARGS$asSLR){
+
+      slr <- as.slr(A)
+      v <- vegan::rda(ARGS$z, slr)
+
+    }else{
+
+      ilr <- compositions::ilr(A)
+      v <- vegan::rda(ARGS$z, ilr)
+    }
+  }, error = function(e){
+
+    v <- numeric() # handle error when SVD fails to converge
+  })
+
+  if(length(v) == 0) return(-1000000)
+  varExplained <- v$CCA$eig / v$tot.chi
+  if(length(varExplained) == 0) return(-1000000) # handle when no CCA inertia
+  return(varExplained)
+}
+
 # objective.maxInvDiag <- function(codon, ARGS){
 #
 # }